+++ packages/generator-teams/src/app/manifestGeneration/ManifestVersions.ts
@@ -7,5 +7,6 @@ export enum ManifestVersions {
     v19 = "v1.9",
     v110 = "v1.10",
     v111 = "v1.11",
-    devPreview = "devPreview"
+    devPreview = "devPreview",
+    m365DevPreview = "m365DevPreview"
 }