+++ packages/yoteams-build-core/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Added telemetry, see https://github.com/pnp/generator-teams/blob/master/docs/docs/about/telemetry.md
 * Added logging of ngrok inspection url (#267 by pschaeflein)
+* Added support for `m365DevPreview` manifest version
 
 ## [*1.4.0*]- <*2021-10-28*>